Good progress on the way to getting TheOther to run
It compiles, but segfaults; I documented the pieces of code
to-be-generated in CodeGen.hs (by hand-fixing the generated C code).

Introduced a Reader monad in some of the Translatable instances, this
allows a context to be routed through, the interface remained stable.

Added a Context.hs module to ease changing the context.

Implemented a way to do 'reasonable' forward declarations; I'm not
happy with how this works (structural duplication!) but it gets the
job done for now and will not leak outside of its module.

Removed lots of stringly typed programming in AST.hs, this uncovered
plenty of bugs (PrettyPrinter.hs printing names as types and
vice-versa, and other stuff like that). Elias will hate me for this,
but it was necessary.