Huge Refactoring of CCode
CCode now is defined as a GADT, which allows to infer which kind of
C-term we're generating without being explicit too much. It now makes
more sense, the types are more fine grained and I found many places
where I was using one of the many type aliases that are/were defined
to be `String` when I should've been using a different one.

I'm NOT happy yet with how I need more constructors than I should be
needing, but I'll find a way to have the cake and eat it, too. (Look
for the `AsExpr . AsLval . Var $ "x"`) pattern to see what I mean.