Added a set-implementation for C
The implementation uses a (unbalanced) binary search tree and supports
adding, removing, subset- and membership-checking, cloning, mapping and
reducing. It has been tested (although not that thoroughly) for
functionality and memory leaks. Let me know if you are missing some
operation. My guess is that this will be clearer when we start
implementing the Par t's.