First version of fully working parser
encorec now accepts as argument a file of encore source code. Note that
an expression using binary operators (+, -, <, >, ==, !=) must be
parenthesised on the outer level. "1+2" is illegal, but (1+2)
and (1+2+3) is allowed. Also note that there are some discrepancies
between the parser and the pretty-printer. This is a todo for later.

Check out mylittlepony/programs for some example programs that parses.