Restructuring of the build system
There is now a Makefile in the root of the repo. *You should use this
Makefile from now on*. It builds everything and creates a folder
"release" containing encorec and two folders "inc" and "lib". They
contain the header and object files needed by the runtime. encorec uses
relative paths to these folders, so only the binary in "release" will
work.

There is probably some additional refactoring to be made to make sure
the Makefiles cleans what they should and do not run rules when they
don't have to.

Note to self: I need to learn more about Makefiles!