Working implementation of closures
There are warnings when using closures that return void. This could
probably be addressed by casts, but I think it would be better if we
could have void-in-encore functions (and methods) return a
representation of the value (), which could be NULL since it should
never be read. It would get rid of a lot of special cases in the code
generation.