Closures can now capture variables
There are no restrictions on what can be captured, so no safety is
guaranteed. The implementation currently uses malloc to set the
environment up (so that it can be freed). Maybe this should be done by
pony_alloc instead some day (as well as the allocation of the closure).

Capturing variables turned out to give some kind of not-so-flexible currying:

class Main
  def main() : void
    let compose = \(g : int -> int, f : int -> int) -> \(x : int) -> g(f(x)) in
      let double = \(x : int) -> x*2 in
        let bump = \(x : int) -> x+1 in
          let f = compose(double, bump) in
            print f(3)