Changed the C-type of bool to int64_t to allow it to be passed as a pony argument
This is of course quite wasteful, space-wise. At some point we should
probably add a smaller bool-type as one of the things that can be
passed as an argument. On the other hand, if it's in a union, it will
probably be as big as an int64_t most of the time anyway...