Temporary fix specific to Linux so far.
My guess for why this fix is necessary is the following:

After `make`, four archives are created in `release/lib`. The output of
`ls` is:  `libclosure.a libfuture.a libpony.a  libset.a` on Linux, with
LANG=en.US.UTF-8. `libfuture.a` depends on `libclosure.a`, so this order
is incorrect when it comes to wildcard marching (using `boolean` as one
example):

    clang -v boolean.pony.c embed_tl.o -ggdb -Wall -lpthread  -o boolean /vagrant/mylittlepony/release/./lib/*.a -I /vagrant/mylittlepony/release/./inc/ -I ..

The weird thing is that it works on XOS without this fix. Future
investigation is needed to reveal the truth.