Support for else-less if statements
else-clauses will bind to the closest if:

if true then
  if false then
    Foo
  else
    Bar

==

if true then
  (if false then
     Foo
   else
     Bar)

Note that if-then is desugared into if-then-else with skip in the else
clause. This means that the then brance in an if-then must have type void.

Also, note that it is possible to use indentation that suggests grouping, but doesn't!