Completed refactoring of the code generation
An encorec program `foo.enc` will be compiled to a separate `.pony.c`
file for each class, plus a header and a file with shared code. These
are placed in the folder `foo_src` (even if the `-o` flag is used to
give the executable a different name).

This fixes the problem where the order of declaration of passive classes
mattered. However, the handling of embedded top level C code is not as
easy since the files that has references to this code have no
declarations to use. So far it works with warnings anyway, but maybe we
should look into it.