Cleaning up build system, use cabal for dependencies.
We saw some problems due to haskell packages missing, having wrong
versions and such. It took a little while to figure out what was
wrong. As we want to be able to specify exactly which libraries we're
using, we (Elias, Stephan) decided to use cabal for that job from now
on. As it turns out, cabal also allows to remove much boring code from
our src/Makefile. I used this opportunity to remove some unused code
from our build system.

This task is part of getting the documentation in shape. After all,
automating dependency management means there is less to document :)