replace future as actors and actors run on its own stack
Replace the current implementation of futures as actors by futures as a
shared data structure. This breaks the thread-safety that we had when using
the previous approach for futures, since the future can be accessed by
multiple actors.

Whenever we create the actor's context, create actor's stack and link it
to the actor's context. It's important to update the field ss_sp after
getting the context since, we update the ss_sp field whenever we getcontext.