Fixed bug in parsing embedded code
Before, you could not write

```
embed void
  future_suspend();
end
```

since the "end" in "suspend" would be parsed as the terminating `end` of
the `embed`. This should now work, with the caveat that empty embeds
fail with non-helpful error messages. This could be fixed by checking
for `try $ do {whiteSpace; reserved "end"}` before parsing any C-code in
the parser, but I think keeping the parser simple trumps this `embed`
"feature".