symlink shared folder to /home/vagrant
This would be the same as sharing the project folder in /home/vagrant.
However, this messes up configuration files such as .profile, .ghc,
etc of the VM since they need to live in /home/vagrant as well.
Therefore, we are going to map the project folder (which is in the
default location, /vagrant) to the default user's home directory.
This change means that after:

  vagrant ssh

you are inside the working directory and there's no need to know the
mapping between the host and the VM (meaning the end user does not need
to type: cd /vagrant)