Fixed bug regarding polymorphism
All type variables were equivalent when looking at the return type of a
method. Now the typechecker uses `hasType e t` when it wants to
typecheck `e` and make sure it has (a subtype of) type `t`, and
`matchTypes t t'` when it wants to see if an expression of type `t'` can
be used as an argument where a type `t` is expected.