Use new argument field names in dispatch function.
The old argument names are `x`,`y`,`bar` -- corresponding to the
encore level names. The new names are `f1`..`fN`. This commit
implements the convention in the dispatch function generation.

To make the produced argument structs more pleasant to read, we also
added an 'Annotated' constructor to the 'CCode' type so the structs
can document the fields.