Add `uint` type.
Literals of the type are written `12u` for now, for easy parsing.

As a future alternative, we could *probably* just emit `IntLiteral`
nodes when a literal is negative and `UIntLiteral` nodes otherwise --
integer coercion should be able handle the rest. But this may also be
somewhat surprising behaviour..

Since this is an urgent feature request, I chose not to do that
now.