Introduce peer keyword
The `peer` keyword is a relative of `new`, with the intended
semantics that a created actor is created in another work pool,
rather than the same as the creating actor.

This is the frontend and code generation, the runtime does not
support this yet. Instead it will print a warning and revert to the
same implementation as `new` for now.