Breathe keyword
The `breathe` keyword makes an actor "come up for air": it will check
whether any thieves are trying to steal work from an actor and
cooperate (if that is the case). The return type is `void`.

The test just checks that nothing crashes, as breathe doesn't have
any visible side effects, nor a return value.