update names of par t types due to name clash with C code
there seems to be a clash of keywords and it's not possible to use bind
and sync as the name of the function. This is due to a bug in the encore
compiler, which does not change the function names of the functions defined
from encore to the generated c code.