creates a virtual environment for encorec
the idea is to create a virtual environment similar to what you
can expect in virtualenv (python) and/or rbenv (ruby). At the moment
it just sets the encore compiler to the right path. Therefore, forking
and having multiple copies of the repo with multiple compilers is
easy to handle. Steps involved:

  1. go to the encore compiler fork you would like to activate
  2. source activate
  3. from this terminal, you can compile any encore program and
     it will safely use the encore compiler that you jut activated.

missing features:

  - no real dependency management yet, so no need to create new folder