task library with futures
first implementation of the task library with futures.
tasks are run by a task runner actor who checks if there
are tasks to execute and, if so, picks up a task from the
mpmcq and runs the task. each `async` returns a future that's
fulfilled whenever the task runner runs its task. The program
doesn't finish until all the tasks are fulfilled!