wait for the tasks to finish when there are no actors available
the current implementation works under mac and linux. however, there is
a problem when doing gc in the future. fulfilling a future fails by
accessing the gc->mark of an actor that seems to be gc'd. i have
disabled gc'ing futures for the time being until i can figure out what
is wrong. added test for tasks with side-effects and when no one is getting
on them.