Make failing tests possible.
Until now it has been possible to test only that something succeeds
(specifically, that the output matches exactly). This limited the
tests to cases where we had determinism and where the expected output
was exactly known.

This commit extends test.sh to maintain the old functionality, but now
ALSO allows checking scripts to be used instead of the old `.out` files.

A checking script is an executable whose name is that of a test, but
with the ending `.chk` -- if the test source if called `foo.enc`, the
checking script is called `foo.chk` and it is executable.

The `make test` command will now compile `foo.enc` and try to run it
(if compilation succeeds). The output of both compilation and
execution will be piped into the checking script. The checking script
can analyse the input and return successfully to signal that
everything is ok, or with a failure code to signal an error.

An example:

File fail.enc:

    class Main
      def main() : void
        repeat i <- 100 {
          print x -- error should occur here
        }

File fail.chk:

    #!/usr/bin/env bash
    stdin=$(cat)

    # Search in stdin for a line that contains "line 4", the error line.
    #
    # If the line is not found, grep returns with failure code,
    # calling the expression after '||', making the test fail:
    echo "$stdin" | grep "line 4" || exit 1

    # To make doubly sure, we also look for the specific message:
    echo "$stdin" | grep "Unbound variable 'x'" || exit 1