Made updating type fields in AST nodes explicit
This commit fixes issue #154. It also makes `setType` in `AST.hs` less
clever in order to avoid similar bugs in the future (the problem was
that `setType` updated any type children to the AST-node, as well as the
meta information -- something that was easy to forget to add when
extending the language. This new version only updates the meta info).

Also, there is now an explicit case in the back-end for typed
expressions (before the type-cast was thrown away after typechecking).