task closure chaining on task-produced futures
implementation of task closures (TASK_CLOSURE). TASK_CLOSURE closures are used
whenever a future produced by a task runner is chained onto another
function. TASK_CLOSURE closures are different from DETACHED closures in the sense
that a detached closure is run by the producer of the future right after
the future is fulfilled while TASK_CLOSURE are packed as messages that will be
run by any task runner (we don't know which one). E.g.

```
  let fut = o.execute() in
    fut ~~> \(x: int) -> x + 1
```

In this case, chaining on the `fut` creates a DETACHED closure on the
lambda. As soon as the `o` object finishes with the `execute` method, it
will check its list of chained closures to run and, run one by one until
there are no more chained closures. This approach is sequential
and, if having 1000 chained closures on the result of `o.execute` it
will take forever.

Let's go for the TASK_CLOSURE closures:

```
  let fut = async { ... } in
    fut ~~> \(x: int) -> x+1
```

In this case, the future produced by the `async` will be fulfilled by
any available task runner (currently, one per scheduler). Furthermore,
after a task  runner finishes running the body of the `async` task, it
doesn't  run all the lambdas chained on the produced future, but creates
a new message and schedules the message to be run by any other task
runner. This means that if we have 1000 chained lambdas on the future,
the task runner needs to place the chained lambdas in the global queue
where any task runner can run the closure. If we have 4 schedulers, we
have 4 task runners that operate in parallel picking up these 1000
closures to run, which load balances the jobs.

Shortcoming: there could be data races if the environment of the
closure captures an passive object. A possible fix to this problem is to
do a copy of the passive object whenever we detect that the closure is impure.