update definitions of the par t types
update some definitions and use of macros when detected repetion of
code. create quasi-universal interface for the new_par
function (couldn't make it universal for the P case, which takes two
arguments instead of one).

TODO: Commented out fmap, need to improve it

Limitations: some cases have not been solved for the general new_par,
such as the FP and M cases.