update functions with no args to pass void type
update functions with no parameters to include `void` argument.
Due to C old list parameter writing style, writing a function (in the
header file)  with no argument means that the number of arguments it's
unknown, and does not throw an error. After updating to the new RT, some
functions change from having one arg to having no arg. This change
prevents this error.