Typechecking with backtracing is now default
Before, the code in `typecheck` would call `pushTypecheck` to push an
expression to the backtrace and typecheck it. Now this same function is
called `typecheck`, and the old `typecheck` is called `doTypecheck`.
This way we get code like this:
```
    doTypecheck fAcc@(FieldAccess {target, name}) = do
      eTarget <- typecheck target
      ...
```
and make all the pushing implicit. A similar refactoring has been done
for `pushHasType`, which has also been moved out from the `Checkable`
type class (since only `Expr` implemented it in practice).