C-level support for upcasts
Before we had subtyping, a type cast (`x : Foo`) was a no-op
dynamically since it could only be used to help the typechecker figure
out the correct type. Now we need to do an actual pointer cast so that
assignments and function argument types get the right types in the
generated code. This PR fixes this.