Possible errors don't need to be advertised in names
Since we are in the typechecker whose sole purpose is to typecheck (and
annotate) the AST or give an error message if it fails, there is no need
to call functions `find_method_or_error`. Almost everything in the
typechecker gives a result or throws an error, and this is visible from
the types (the `ExceptT` monad transformer).