add option types to encore
addition of option types to encore. there is a type constructor called `Maybe t`
which represent the option type. valid data constructors for option
types are `Just t` and `Nothing`. there are some examples in the file
`maybeType.enc`. a small example would be:

    def test_catch_all(z: Maybe int): string
      match z with
      z => "catchall"
      Just 32 => "Error, general condition should have been catched
      first"

this is translated to c into a bunch of if-else statements. i mean a
bunch because the developer can match on nested option types of the form
`Maybe (Maybe int)`.

restrictions:
  - there is not match all clause in the `match` expression