Major refactoring of type implementation
This commit refactors the implementation of types. Most changes are in
`Types.hs`, but many of these changes have propagated to other places.
There are some other refactorings done to related code, mostly to unify
conventions and fix minor bugs found along the way. This refactoring
also removes the need to disable certain GHC-warnings (which should
always be a warning that you should implement things differently).

When parsing a reference type, the resulting type is an `UntypedRef`,
which during typechecking is converted into a `ClassType` or a
`TraitType`, separating the overloaded use of the old `RefType`. Only
`ClassType` has activity and capability information (see below), and
this design allows adding or removing information to one kind of
`RefType` without affecting the representation of the other.

To resolve lookups to the correct type, a new phase has been added to
the compiler (in `Prechecker.hs`). It visits all classes, traits and
functions and typechecks their interfaces (up to but not including
expressions). This way, we know that looking up a field/method/function
will never result in a type containing an `UntypedRef`. Also, it reduces
the complexity of `Typechecker.hs` which only needs to perform
typechecking of expressions and check that the included traits of
classes are well formed. The shared functions have been moved to
the new module `Typechecker.Util`.

There is a new `CapabilityType` that stores a `Capability`. This is
meant to store the representation of composite capabilities like `(Foo +
Bar) * Fr0b` when it is implemented, but is currently just a list of
trait types. As an example, if a class is declared as
```
passive class C : Foo + Bar<int> + Baz{...}
```
the type `C` has the capability `[Foo, Bar<int>, Baz]`. When there is
proper support for composite capabilities, the representation of
`Capability` should probably change to something more tree-like, but
hopefully this refactoring will make the transition smoother by
abstracting over the representation.

A big change from before is that `RefType`s no longer store the complete
declarations of their included traits, but rather just enough
information to lookup the needed information from the environment. The
motivation for doing this is to avoid duplication of information, so
that all information about a certain trait is always found in the same
place, and if some later compiler pass needs to change a trait, there is
no confusion to where this change should be done. This changes how
translation is handled, but also removes the need for the AST-node
`ImplementedTrait`.

This commit seems to fix #190.