Experimental support for borrowing
This commit adds support for declaring parameters as
stackbound (`S(Foo)`). A linear reference can be passed as a stack bound
parameter without consuming, as the linear alias is buried for the
duration of the call (borrowing requires a synchronous call). A
stackbound may not be aliased (except through borrowing), and cannot be
used in parallel (so far this means that it may not be passed to an
active object). The following program exemplifies most uses of
borrowing:

```
def borrow(x : S(Token)) : void{
  let thief = new Thief<Token>
      -- y = consume x -- Stackbounds cannot be bound to new variables
  in{
    -- thief.steal(x); -- Stackbounds cannot be passed to active objects
    x.foo(); -- You can call methods on stackbound objects
  }
}

trait Foo
 def foo() : void
   print "Foo!"

passive class Token : linear Foo

class Thief<t>
  def steal(x : S(t)) : void
    ()

class Main
  -- f : S(Foo<int>) -- Fields cannot be stackbound
  def borrow(x : S(Token)) : void
    borrow(x) -- Stackbounds can be passed to other borrowing functions/methods

  def main() : void
    let x = new Token
        thief = new Thief<Token>
    in{
       -- thief.steal(x); -- Linear values cannot be borrowed by active objects
       this.borrow(x); -- Synchronous method calls can borrow linear values
    }
```

Borrowing is so far very conservative with respect to what kind of
aliasing is allowed. Programming with it should show where additional
programming conveniences is needed.