refactor AST to work with bottom type
refactor AST of MaybeData to work with bottom type. before, we had a
meta at the MaybeData node and at the kind of MaybeData, represented by
a JUST or NOTHING. This is redundant and caused the typechecker to
update the metadata of the MaybeData to some type (by inference) but not
update the JUST / NOTHING node with any metadata. This means that in
case of NOTHING, the metadata would still have type Bottom, which we do
not want to pass to the backend.