fix typechecking option types and match statements
the option type `Nothing` is cast to a `bottom` type. the backend
doesn't know how to deal with `bottom` type and therefore, if there is
any `bottom` type when using `Nothing` it should throw an error.

fix `match` statement with a bare-minimum of type inference on the
pattern match clause. for instance:

    match Just Just 32 with {
      Just Nothing => ...
    }

infers that the type of Nothing should be `Maybe int`. therefore, it
replaces `Just Nothing` with previous type `Maybe Maybe Bottom` by the
type `Maybe Maybe int`.