Support for `val` fields
This commit adds support for declaring read-only fields. Such a field is
declared using the keyword `val` and may not be written to, except in
constructors. This allows us to have covariance on field types for
included traits:
```
trait HasPet
  require val pet : Animal

passive class DogPerson : HasPet
  pet : Dog

passive class CatPerson
  pet : Cat
```
(where `Dog <: Animal` and `Cat <: Animal`)

Since the methods of `HasPet` cannot update the field `pet`, the actual
type of `pet` can be any subtype of `Animal`, which increases trait
reuse. `val`-fields are also needed for `readonly` capabilities. Having
`val`-fields in a class means that any updates of the field must happen
in the constructor, or in one of the included methods (that does not
have a "`val` view" of the trait). See the added test for a slightly
more realistic usage.

Minor fixes in this commit is an update of the keywords in
`encore-mode.el`, insertion of C-level type casts for binary
operators (see #208) and an update of the cabal dependencies to support
`base 4.8.1.0`.