refactor test and fix infinite loop in match expr
fix infinite loop of a match expression when the variable to be bound
was of the form:

    match Just(32) with
      z => ...    -- this used to get stuck in an infinite loop

this used to get stuck in an infinite loop when doing the
typechecking. it was fixed by adding a new case that knows how to deal
with items of the AST node type `VarAccess`.

refactor tests to be more descriptive and maintanable