Adding support for for loops and simple ranges
    for i <- [1..5] print("{} ", i)

prints "1 2 3 4 5". If my_array = [1,2,3,4,5],

    for i <- my_array print("{} ", i)

prints "1 2 3 4 5". The following all print "1 3 5":

    for i <- my_array by 2 print("{} ", i)
    for i <- [1..5] by 2 print("{} ", i)
    for i <- [1..5 by 2] print("{} ", i)

The trailing by on the foreloop is useful for changing
the stride length when you are iterating over an array
or a range that weren't given as a literal. They do
combine, so 'the crazy' statement:

    for i <- [1..5 by 2] by 2 print("{} ", i)

prints "1 5".

Added a test for for loops