First draft working version of Strings
TopLevel.hs calls addStdLib which adds an import String.enc (For now) to
the top-level file.

There is also some refactoring of how print statements are handled, and
with it some not very nice additions to the typechecker. I believe this
can be refactored later. The problem was that `print("The answer is
{}\n", 42)` was parsed as function call, but desugaring into a `Print`
node only worked when the first argument was a string literal. But we
had already desugared the literal into a `String` object, so the print
desugaring didn't fire. In the end I think we actually have better error
messages, but the code could probably be improved.