fix precedence, example and naming convention
fix the precedence of the `liftf` and `liftv` combinators, they have
higher precedence than the `||` combinator. re-wrote the example to make
it more understandable while, at the same time, testing all
aspects. renaming of `par_s` to `par_t` to keep code style similar to
other modules. fix typechecking rules for `>>` and `||` to check that
they have to be `par` types.