Added support for warnings during typechecking
During the typechecking phase you can now emit warnings which will be
printed at compile time. Warnings are not arbitrary strings (like
errors) but must be explicitly defined as constructors of the type
`Warning`, together with the corresponding case for showing the warning.
Right now there is only one warning `StringDeprecatedWarning`, which is
given when the type `string` is used. It is produced in `Util.hs` in the
following code:

```
| isStringType ty = do
    tcWarning StringDeprecatedWarning
    return ty
```

I believe we should move to this system for errors as well as it
centralizes all possible errors that can be thrown. For example, we
would then write `assertSubtypeOf` (in `Typechecker.hs`) as

```
assertSubtypeOf :: Type -> Type -> TypecheckM ()
assertSubtypeOf sub super =
    unlessM (sub `subtypeOf` super) $
            tcError $ TypeMismatch sub super
```