comment out things that are not used currently
comment out the parts of the code that are not reachable currently. the
only part of the combinators that cannot be tested at the moment is the
empty combinator. partially, comment out the JOIN_PAR case which is not
used yet in some scenarios. As soon as we can create / tag something
with the JOIN_PAR, these could be re-added.