add each combinator to encore compiler
this commit adds support for the `each` combinator. this combinator is
useful when we want to take an array of items (at the Encore level) and
be able to perform parallel computations on those items.

TODO:
- the party library will split the array in lists of 1000 items. any
operation performed on the items will be handled by a task. at the
moment, if we have 2000 items there will be two tasks in charge of
handling those 2000 items, 1000 items for each one. the runtime is not
smart enough to:

  (each [1..2000]) >> long_computation

be able to do any load-balancing and/or split the array in smaller
tasks. the implication of this is that the task will perform the
computation on the 1000 elements without giving up his thread of control
until it has finished.