Allow composite traits to be used as types
There was no support in the backend for using composite traits as types,
causing programs like the following to crash.
```
trait T
trait U

passive class C : T + U -- This is fine, not used as a type

class Main
  def main() : void
    new C : T + U -- This crashes the compiler
```

This commit adds proper backend support for types like `T + U`,
introducing the C type `capability`, which is a pointer to a struct that
contains only the self type that prefixes all passive classes. This
means that any reference to a passive class (trait and capability
references included) can be cast to a `capability` and used as such.

This commit also adds proper tracing support for trait types, which were
erroneously ignored previously during tracing. It should therefore now
be safe to have fields of trait type or to pass trait references between
active objects.