Added syntactic sugar for variable introduction
This commit adds the following rule to the desugaring stage:

```
{
  let x = e1;
  e2;
  ...;
  en;
}

=>

{
  let x = e1 in {e2; ...; en}
}
```

This lets you add variables in a scope (seemingly) without introducing a
nested scope. The rule works left to right, so if there are more `let`s
in `e2; ...; en`, these will also be unfolded. An example of what you
can write now is in the test `miniLet.enc`.