add parametric functions to encore
add parametric functions to encore. this commits fixes the tracing of
the runtime type of parametric functions. an example to prove this is
provided, although it needs to be checked manually, as the tracing
function is something that happens at runtime. an example of a
parametric function is:

```
def <t> createNode(content: t): Node {
  new Node<t>(content)
}

class Node<t>
  ...

  def init(content: t): void
    ...
```