Refactor typechecker to hide type substitution
This commit refactors the way type substitution is handled so that it is
completely encapsulated in the lookup rules of the environment. As an
example, take the following class
```
passive class Foo<t>
  f : t

class Main
  def main() : void {
    let x = new Foo<int>;
    x.f = 42;
  }
```

In order to ensure that the assignment on the last line is valid, we
need to build a set of bindings that maps `t` to `int`, so that the
field `f : t` looks like `f : int` when we do the lookup. Instead of
cluttering the typechecker with this sort of thing, this translation is
now done implicitly on field lookup.