Add tracing of embedded data
This commit adds support for tracing embedded data. Since there is no
way of knowing how this data looks, it is assumed to be a single block
of memory with no pointers. Users who want to embed linked data need to
define their own tracing functions and call it the wrapping class:

```
class Foo
  f : embed linked_t* end

  def Foo_trace() : void {
    let f = this.f;
    embed void
      my_trace_function(_ctx, #{f});
    end
  }
```

The test `stringTrace.enc` will trigger a garbage collection in `Foo`
that earlier caused one of the strings' embedded data to be prematurely
collected.