fix #372: misleading build instructions
there was missing information on how to install Encore when you already
have a different version of `ghc`. this commit adds information on how to
make it work with different versions. the solution is explained by
updating the `$PATH` variable to start with `/opt/ghc`. the commit
provides information on how to do this globally or locally.