Let pretty printing use OverloadedStrings
This commit refactors the pretty printing of C and Encore ASTs, mainly
by using [OverloadedStrings](https://www.schoolofhaskell.com/school/to-infinity-and-beyond/pick-of-the-week/guide-to-ghc-extensions/basic-syntax-extensions#overloadedstrings). This GHC extension lets you use string literals
where another `String` like data-type is expected. The pretty printers
work with the type `Doc`, meaning we had to write `text "*"` when we
wanted a literal `"*"`. Now, the string literal is enough. Also cleaned up
parts of the code and made the emacs mode more resilent to change.