fix fut chaining runtime acquiring more references
problem: the future that performs the chain may receive an object for
which it has not acquire any references. Therefore, it should acquire
references. In case of the object having already references or being the
owner of the object, the action of acquiring more references will be
ignored. So, it is safe to acquire more references even when you own the
object (own object = lives in your heap). this commit fixes #412.