Separate resolveType and checking for recursion (#408)
* Separate resolveType and checking for recursion

This commit refactors `resolveType` so that the checking for recursive
type synonyms is not done for every lookup. This allows it to be lifted
out of the state monad (which is the main reason for this refactoring).
Instead there is a function `resolveTypeAndCheckForLoops` that does what
`resolveType` did before, and this function is called when prechecking a
type synonym. This is done with minimal code duplication by also
refactoring resolving of ref types into its own function.

* Added new test case.

* Updated .gitignore