Support for double barring
This commit changes how residual aliasing is handled.

`CAT(x.f, y, y.g) => z` now requires `z` to be a field in the type of
`x.f`. Additionally, `z` must be *strongly restricted* in the type of
`x.f`, denoted by `|| z`. Strong restrictions are similar to weak
restrictions (single barring), but assigning to a field with a strong
restriction requires that the value being assigned is unrestricted. This
means that ownership transfer to a strongly restricted field means that
some field ownership is lost, unless its retained through a residual
alias. The invariant if `x : T || f` is that there is no alias of `x`
that has access to `f`.