refactor tests to not-used `extract`
refactor tests in ParT to avoid the use of the `extract` function. this
is due to an issue (#434) related to futures. the sequence test omits
the use of extract, and the liftf test has been re-written to avoid that
as well. tests that cannot avoid the use of `extract` have been disabled
and will be re-enable as soon as this works.